Update the script with compacted output and remove Bash trap TODO
I made the script run every 10 seconds and expire the status in a minute. This reduces the need to force-update the status as it happens frequently now, so I removed the Enter trap TODO. I have also documented the problem with artist and song desync but since it fixes itself in 10 seconds now I no longer consider it worth fixing.